+++ src/precis/data/templates/draft/__init__.py
@@ -1,0 +1,2 @@
+""
+"""LaTeX draft-export template assets."""